Return same SafeInfo from Blockchain and DB
- Closes #864
